Move DomTester into .test
