+++ tests/src/test/scala/japgolly/scalajs/react/core/HooksTest.scala
@@ -3,6 +3,7 @@ package japgolly.scalajs.react.core
 import japgolly.scalajs.react.Hooks.UseEffectArg
 import japgolly.scalajs.react._
 import japgolly.scalajs.react.extra._
+import japgolly.scalajs.react.test.DomTester
 import japgolly.scalajs.react.test.ReactTestUtils._
 import japgolly.scalajs.react.test.TestUtil._
 import japgolly.scalajs.react.vdom.html_<^._
